RF_02.001 | FreestyleProjectConfigurationTest https://github.com/RedRoverSchool/JenkinsQA_Java_2025_spring/issues/1644 ER_02.001 | FreestyleProjectConfigurationTest https://github.com/RedRoverSchool/JenkinsQA_Java_2025_spring/issues/1655
